refactoring to separate field
